// src/generated/models/view-action-definition-model.ts
import type { ViewDefinition } from './view-definition-model';

export const ViewActionDefinitionHandlerKeyKeyToLabel = {
  'AddComment': 'add_comment',
  'AddAttachment': 'add_attachment',
  'EditRequest': 'edit_request',
  'TransitionRequest': 'transition_request'
} as const;
export type ViewActionDefinitionHandlerKeyKey = keyof typeof ViewActionDefinitionHandlerKeyKeyToLabel;

export const ViewActionDefinitionPlacementKeyToLabel = {
  'Header': 'Header',
  'Section': 'Section'
} as const;
export type ViewActionDefinitionPlacementKey = keyof typeof ViewActionDefinitionPlacementKeyToLabel;

export const ViewActionDefinitionStatusKeyToLabel = {
  'Active': 'Active',
  'Inactive': 'Inactive'
} as const;
export type ViewActionDefinitionStatusKey = keyof typeof ViewActionDefinitionStatusKeyToLabel;

export interface ViewActionDefinition {
  /**
   * @displayName View Action Definition ID
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Action Name
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  actionName: string;
  /**
   * @displayName Action Key
   * @validationRule Required for create/update operations
   */
  actionKey: string;
  /**
   * @displayName Handler Key
   * @validationRule Required for create/update operations
   */
  handlerKeyKey: ViewActionDefinitionHandlerKeyKey;
  /**
   * @displayName Placement
   * @validationRule Required for create/update operations
   */
  placementKey: ViewActionDefinitionPlacementKey;
  /**
   * @displayName Required Permission
   */
  requiredPermission?: string;
  /**
   * @displayName Section Key
   */
  sectionKey?: string;
  /**
   * @displayName Sort Order
   * @validationRule Required for create/update operations
   */
  sortOrder: number;
  /**
   * @displayName Status
   * @validationRule Required for create/update operations
   */
  statusKey: ViewActionDefinitionStatusKey;
  /**
   * @displayName View Definition
   * @validationRule Required for create/update operations
   */
  viewDefinition: Pick<ViewDefinition, 'id' | 'viewName'>;
}

export const _ViewActionDefinition = 'ViewActionDefinition' as const;